requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
alembic==1.0.7            # via flask-migrate
bcrypt==3.1.6             # via flask-bcrypt
certifi==2018.11.29       # via requests
cffi==1.12.1              # via bcrypt
chardet==3.0.4            # via requests
click==7.0                # via flask
faker==1.0.2
flask-bcrypt==0.7.1
flask-cors==3.0.7
flask-migrate==2.4.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
gunicorn==19.9.0
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
mako==1.0.7               # via alembic
markupsafe==1.1.0         # via jinja2, mako
pycparser==2.19           # via cffi
pyjwt==1.7.1
pymysql==0.9.3
python-dateutil==2.8.0    # via alembic, faker
python-editor==1.0.4      # via alembic
requests==2.21.0
six==1.12.0               # via bcrypt, faker, flask-cors, python-dateutil
sqlalchemy==1.2.18        # via alembic, flask-sqlalchemy
text-unidecode==1.2       # via faker
urllib3==1.24.1           # via requests
werkzeug==0.14.1          # via flask
wtforms==2.2.1            # via flask-wtf
